+++ tests/asyncio/test_connection.py
@@ -205,3 +205,23 @@ async def test_run_visitor_exception(asyncio_engine, mytable):
     async with asyncio_engine.connect() as conn:
         await conn._worker.run(thread_fn, conn)
         assert thread_called
+
+
+@pytest.mark.asyncio
+async def test_sync_cm_exception(asyncio_engine):
+    thread_called = False
+
+    def thread_fn(conn):
+        nonlocal thread_called
+
+        meta = MetaData()
+        with pytest.raises(TypeError, match='Use async with'):
+            meta.reflect(conn)
+
+        meta.reflect(conn.sync_connection)
+
+        thread_called = True
+
+    async with asyncio_engine.connect() as conn:
+        await conn._worker.run(thread_fn, conn)
+        assert thread_called